Modified to remove MS warnings d:\coin\coinall-trunk\osi\src\osibranchingobject.cpp(523) : warning C4244: '=' : conversion from 'double' to 'float', possible loss of data

+++ trunk/Osi/src/OsiBranchingObject.hpp
@@ -220,7 +220,7 @@ protected:
   /// Preferred way of branching - -1 off, 0 down, 1 up (for 2-way)
   int preferredWay_;
   /// "Infeasibility" on other way
-  mutable float otherInfeasibility_;
+  mutable double otherInfeasibility_;
   
 };
